+++ ecs/query_no_debug.go
@@ -1,0 +1,9 @@
+//go:build !debug
+
+package ecs
+
+const isDebug = false
+
+func (q *Query) checkNext() {}
+
+func (q *Query) checkGet() {}